[STYLE] rename post file
